+++ CHANGELOG.md
@@ -7,7 +7,7 @@ This document is user facing. Please word the changes in such a way
 that users understand how the changes affect the new version.
 --->
 
-## v0.1.4-dev
+## v0.1.4
 + Add support for drawing only part of the transcript
 + Add support for gene names (defaults to MANE_Select transcript)
 + Add option to show exon number